move streaming parser into its own submodule
